inital tableview refactor
