+++ HackIllinois/Controllers/ViewControllers/HIAnnouncmentsViewController.swift
@@ -11,7 +11,7 @@ import UIKit
 import CoreData
 
 class HIAnnouncmentsViewController: HIBaseViewController {
-
+    // MARK - Properties
     lazy var fetchedResultsController: NSFetchedResultsController<Announcement> = {
         let fetchRequest: NSFetchRequest<Announcement> = Announcement.fetchRequest()
 
@@ -23,41 +23,30 @@ class HIAnnouncmentsViewController: HIBaseViewController {
 
         return fetchedResultsController
     }()
-
 }
 
-// MARK: - IBActions
+// MARK: - Actions
 extension HIAnnouncmentsViewController {
 
 }
 
 // MARK: - UIViewController
 extension HIAnnouncmentsViewController {
-
-//    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-//        super.prepare(for: segue, sender: sender)
-//        if let indexPath = sender as? IndexPath, let eventDetailViewController = segue.destination as? HIEventDetailViewController {
-//            eventDetailViewController.model = fetchedResultsController.object(at: indexPath)
-//        }
-//    }
-
     override func viewDidLoad() {
         super.viewDidLoad()
         _fetchedResultsController = fetchedResultsController as? NSFetchedResultsController<NSManagedObject>
-        collectionView?.register(UINib(nibName: HIAnnouncementCell.IDENTIFIER, bundle: nil), forCellWithReuseIdentifier: HIAnnouncementCell.IDENTIFIER)
-//        collectionView?.register(UINib(nibName: HIDateHeader.IDENTIFIER, bundle: nil), forSupplementaryViewOfKind: UICollectionElementKindSectionHeader, withReuseIdentifier: HIDateHeader.IDENTIFIER)
+        tableView?.register(UINib(nibName: HIAnnouncementCell.IDENTIFIER, bundle: nil), forCellReuseIdentifier: HIAnnouncementCell.IDENTIFIER)
 
-        collectionView?.contentInset = UIEdgeInsets(top: 0, left: 12, bottom: 0, right: 12)
+        tableView?.contentInset = UIEdgeInsets(top: 0, left: 12, bottom: 0, right: 12)
 
         try! fetchedResultsController.performFetch()
     }
-
 }
 
-// MARK: - UICollectionViewDataSource
+// MARK: - UITableViewDataSource
 extension HIAnnouncmentsViewController {
-    override func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: HIAnnouncementCell.IDENTIFIER, for: indexPath)
+    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        let cell = tableView.dequeueReusableCell(withIdentifier: HIAnnouncementCell.IDENTIFIER, for: indexPath)
         if let cell = cell as? HIAnnouncementCell {
             //            cell.titleLabel.text = fetchedResultsController.object(at: indexPath).name
             cell.titleLabel.text = "announcement"
@@ -65,10 +54,4 @@ extension HIAnnouncmentsViewController {
         return cell
     }
 }
-
-// MARK: - UICollectionViewDelegateFlowLayout
-extension HIAnnouncmentsViewController {
-
-}
-
