--- HackIllinois/Controllers/ViewControllers/HIBaseViewController/HIBaseViewController+AlertPresentation.swift
@@ -1,26 +1,0 @@
-//
-//  HIBaseViewController+AlertPresentation.swift
-//  HackIllinois
-//
-//  Created by Rauhul Varma on 11/21/17.
-//  Copyright © 2017 HackIllinois. All rights reserved.
-//
-
-import Foundation
-import UIKit
-
-extension HIBaseViewController {
-
-    func presentErrorController(title: String, message: String? = nil, dismissParentOnCompletion exit: Bool) {
-        let alert = UIAlertController(title: title, message: message, preferredStyle: .alert)
-        alert.addAction(
-            UIAlertAction(title: "Ok", style: .default) { (_) in
-                if exit {
-                    self.dismiss(animated: true, completion: nil)
-                }
-            }
-        )
-        present(alert, animated: true, completion: nil)
-    }
-
-}
