--- HackIllinois/Controllers/ViewControllers/HIBaseViewController/HIBaseViewController+HICollectionViewDelegateFlowLayout.swift
@@ -1,11 +1,0 @@
-//
-//  HIBaseViewController+HICollectionViewDelegateFlowLayout.swift
-//  HackIllinois
-//
-//  Created by Rauhul Varma on 11/23/17.
-//  Copyright © 2017 HackIllinois. All rights reserved.
-//
-
-import Foundation
-
-extension HIBaseViewController: HICollectionViewDelegateFlowLayout { }
