--- HackIllinois/Controllers/ViewControllers/HIBaseViewController/HIBaseViewController+KeyboardNotifications.swift
@@ -1,40 +1,0 @@
-//
-//  HIBaseViewController+KeyboardNotifications.swift
-//  HackIllinois
-//
-//  Created by Rauhul Varma on 11/20/17.
-//  Copyright © 2017 HackIllinois. All rights reserved.
-//
-
-import Foundation
-import UIKit
-
-extension HIBaseViewController {
-
-    func registerForKeyboardNotifications() {
-        NotificationCenter.default.addObserver(self, selector: #selector(keyboardWillShow(_:)), name: .UIKeyboardWillShow, object: nil)
-        NotificationCenter.default.addObserver(self, selector: #selector(keyboardWillHide(_:)), name: .UIKeyboardWillHide, object: nil)
-    }
-
-    func unregisterForKeyboardNotifications() {
-        NotificationCenter.default.removeObserver(self, name: NSNotification.Name.UIKeyboardWillShow, object: nil)
-        NotificationCenter.default.removeObserver(self, name: NSNotification.Name.UIKeyboardWillHide, object: nil)
-    }
-
-    func animateWithKeyboardLayout(notification: NSNotification, layout: ((CGRect) -> Void)?) {
-        guard let duration = notification.userInfo?[UIKeyboardAnimationDurationUserInfoKey] as? Double,
-            let curveValue = notification.userInfo?[UIKeyboardAnimationCurveUserInfoKey] as? Int,
-            let curve = UIViewAnimationCurve(rawValue: curveValue),
-            let keyboardFrameValue = notification.userInfo?[UIKeyboardFrameEndUserInfoKey] as? NSValue else { return }
-
-        let keyboardFrame = keyboardFrameValue.cgRectValue
-
-        layout?(keyboardFrame)
-
-        UIView.beginAnimations(nil, context: nil)
-        UIView.setAnimationDuration(duration)
-        UIView.setAnimationCurve(curve)
-        view.layoutIfNeeded()
-        UIView.commitAnimations()
-    }
-}
